Deduplicated fabric elements types
