+++ src/fabric/RadialGradientNativeComponent.ts
@@ -1,25 +1,7 @@
 import codegenNativeComponent from 'react-native/Libraries/Utilities/codegenNativeComponent';
-import type {
-  Float,
-  Int32,
-  WithDefault,
-} from 'react-native/Libraries/Types/CodegenTypes';
+import type { Float, Int32 } from 'react-native/Libraries/Types/CodegenTypes';
 import type { ViewProps } from './utils';
-
-interface SvgNodeCommonProps {
-  name?: string;
-  opacity?: WithDefault<Float, 1.0>;
-  matrix?: ReadonlyArray<Float>;
-  mask?: string;
-  markerStart?: string;
-  markerMid?: string;
-  markerEnd?: string;
-  clipPath?: string;
-  clipRule?: WithDefault<Int32, 0>;
-  responsible?: boolean;
-  display?: string;
-  pointerEvents?: string;
-}
+import type { SvgNodeCommonProps } from './commonTypes';
 
 interface NativeProps extends ViewProps, SvgNodeCommonProps {
   fx?: string;
