+++ src/fabric/commonTypes.ts
@@ -1,0 +1,80 @@
+import type {
+  Float,
+  Int32,
+  WithDefault,
+} from 'react-native/Libraries/Types/CodegenTypes';
+import type { ColorValue } from 'react-native';
+
+export interface SvgNodeCommonProps {
+  name?: string;
+  opacity?: WithDefault<Float, 1.0>;
+  matrix?: ReadonlyArray<Float>;
+  mask?: string;
+  markerStart?: string;
+  markerMid?: string;
+  markerEnd?: string;
+  clipPath?: string;
+  clipRule?: WithDefault<Int32, 0>;
+  responsible?: boolean;
+  display?: string;
+  pointerEvents?: string;
+}
+
+type ColorStruct = Readonly<{
+  type?: WithDefault<Int32, -1>;
+  payload?: ColorValue;
+  brushRef?: string;
+}>;
+
+export interface SvgRenderableCommonProps {
+  fill?: ColorStruct;
+  fillOpacity?: WithDefault<Float, 1.0>;
+  fillRule?: WithDefault<Int32, 1>;
+  stroke?: ColorStruct;
+  strokeOpacity?: WithDefault<Float, 1.0>;
+  strokeWidth?: WithDefault<string, '1'>;
+  strokeLinecap?: WithDefault<Int32, 0>;
+  strokeLinejoin?: WithDefault<Int32, 0>;
+  strokeDasharray?: ReadonlyArray<string>;
+  strokeDashoffset?: Float;
+  strokeMiterlimit?: Float;
+  vectorEffect?: WithDefault<Int32, 0>;
+  propList?: ReadonlyArray<string>;
+}
+
+type FontObject = Readonly<{
+  fontStyle?: string;
+  fontVariant?: string;
+  fontWeight?: string;
+  fontStretch?: string;
+  fontSize?: string;
+  fontFamily?: string;
+  textAnchor?: string;
+  textDecoration?: string;
+  letterSpacing?: string;
+  wordSpacing?: string;
+  kerning?: string;
+  fontFeatureSettings?: string;
+  fontVariantLigatures?: string;
+  fontVariationSettings?: string;
+}>;
+
+export interface SvgGroupCommonProps {
+  fontSize?: string;
+  fontWeight?: string;
+  font?: FontObject;
+}
+
+export interface SvgTextCommonProps {
+  dx?: ReadonlyArray<string>;
+  dy?: ReadonlyArray<string>;
+  x?: ReadonlyArray<string>;
+  y?: ReadonlyArray<string>;
+  rotate?: ReadonlyArray<string>;
+  inlineSize?: string;
+  textLength?: string;
+  baselineShift?: string;
+  lengthAdjust?: string;
+  alignmentBaseline?: string;
+  verticalAlign?: string;
+}
